Moved EventEditActivity to separate package
